move some common code
